util/linuxfw: add improved health message for bad kernels
Based on recent bug activity it appears a good number of major distros
have just shipped Linux 6.8.0, a no longer maintained branch, with
backports from other branches but not their corresponding fixes.

Add an explicit check for this output, including a pointer to the
relevant kernel commit, and the suggested workaround.

Updates #13863

Signed-off-by: James Tucker <james@tailscale.com>